MOVEONLY: Move NodeContext struct to node/context.h
